added sqrt expr and fixed dot product
